Cleanup migration of sources and device display from playback to system
