+++ src/components/features/PlaylistScreen.tsx
@@ -32,7 +32,7 @@ const PlaylistScreen: FC = () => {
         SCREEN_HEADER_HEIGHT,
         SCROLL_POS_DISPATCH_RATE,
     } = useAppGlobals();
-    const currentSource = useAppSelector((state: RootState) => state.playback.current_audio_source);
+    const currentSource = useAppSelector((state: RootState) => state.system.streamer.sources?.active);
     const playStatus = useAppSelector((state: RootState) => state.playback.play_status);
     const { power: streamerPower } = useAppSelector((state: RootState) => state.system.streamer);
     const playlistViewportRef = useRef<HTMLDivElement>(null);
